chore(test): extract motion.rs test module into motion/tests.rs
Final consistency pass: motion.rs (~1256 lines, ~800 tests) → motion/mod.rs
(production: MotionState, WanderPhase, advance_wander, WalkPathSnapshot) +
motion/tests.rs as a `#[cfg(test)] mod tests;` submodule (`use super::*`,
full crate-internal access, no API change).

All three large tui modules now share the layout: production in mod.rs, tests
in a sibling #[cfg(test)] submodule.

561 tests, clippy -D warnings + fmt clean.
